+++ Library/include/CSP/Systems/Settings/ApplicationSettings.h
@@ -60,7 +60,7 @@ public:
 private:
     ApplicationSettingsResult(void*) {};
 
-    void OnResponse(const csp::services::ApiResponseBase* ApiResponse) override;
+    CSP_NO_EXPORT void OnResponse(const csp::services::ApiResponseBase* ApiResponse) override;
 
     csp::common::ApplicationSettings ApplicationSettings;
 };