+++ Library/src/Systems/Users/UserSystem.cpp
@@ -71,21 +71,25 @@ void StartMultiplayerConnection(csp::multiplayer::MultiplayerConnection& Multipl
     }
 }
 
-/* Check if the provided expiry length in token options is formatted as "HH:MM:SS"
+/* Check if the provided expiry length in token options is formatted as "HH:MM:SS" or "HHH:MM:SS"
  *
- * Return True if expiry length matches format "HH:MM:SS" or empty, false otherwise */
+ * Return True if expiry length matches format "HH:MM:SS" or "HHH:MM:SS", false otherwise
+ *
+ * "HHH:MM:SS" supports durations greater than 4 days */
 bool CheckExpiryLengthFormat(const csp::common::String& ExpiryLength)
 {
-    std::regex Regex("^[0-9]{2}:[0-9]{2}:[0-9]{2}$");
-    if (!ExpiryLength.IsEmpty() && std::regex_search(ExpiryLength.c_str(), Regex))
+    if (ExpiryLength.IsEmpty())
+    {
+        return false;
+    }
+
+    std::regex Regex("^[0-9]{2,3}:[0-5][0-9]:[0-5][0-9]$");
+    if (std::regex_search(ExpiryLength.c_str(), Regex))
     {
         return true;
     }
-    else
-    {
-        CSP_LOG_MSG(csp::common::LogLevel::Warning, "Expiry length token option does not match the expected format, and has been ignored.");
-    }
 
+    CSP_LOG_MSG(csp::common::LogLevel::Warning, "Expiry length token option does not match the expected format, and has been ignored.");
     return false;
 }
 
@@ -145,6 +149,11 @@ void AuthContext::RefreshToken(std::function<void(bool)> Callback)
         Request->SetDeviceId(csp::CSPFoundation::GetDeviceId());
         Request->SetUserId(LoginState->UserId);
         Request->SetRefreshToken(LoginState->RefreshToken);
+
+        auto Options = std::make_shared<chs_user::TokenOptions>();
+        Options->SetExpiryLength(LoginState->AccessTokenExpiryLength);
+        Options->SetRefreshTokenExpiryLength(LoginState->RefreshTokenExpiryLength);
+        Request->SetTokenOptions(Options);
 
         LoginStateResultCallback LoginStateResCallback = [=](const LoginStateResult& LoginStateRes)
         {
@@ -245,12 +254,21 @@ void UserSystem::Login(const csp::common::String& UserName, const csp::common::S
             Request->SetVerifiedAgeEighteen(*UserHasVerifiedAge);
         }
 
-        if (TokenOptions.HasValue() && CheckExpiryLengthFormat(TokenOptions->ExpiryLength))
+        auto Options = std::make_shared<chs_user::TokenOptions>();
+
+        if (TokenOptions.HasValue() && CheckExpiryLengthFormat(TokenOptions->AccessTokenExpiryLength))
         {
-            auto Options = std::make_shared<chs_user::TokenOptions>();
-            Options->SetExpiryLength(TokenOptions->ExpiryLength);
-            Request->SetTokenOptions(Options);
+            Options->SetExpiryLength(TokenOptions->AccessTokenExpiryLength);
+            CurrentLoginState.AccessTokenExpiryLength = TokenOptions->AccessTokenExpiryLength;
         }
+
+        if (TokenOptions.HasValue() && CheckExpiryLengthFormat(TokenOptions->RefreshTokenExpiryLength))
+        {
+            Options->SetRefreshTokenExpiryLength(TokenOptions->RefreshTokenExpiryLength);
+            CurrentLoginState.RefreshTokenExpiryLength = TokenOptions->RefreshTokenExpiryLength;
+        }
+
+        Request->SetTokenOptions(Options);
 
         LoginStateResultCallback LoginStateResCallback = [=](const LoginStateResult& LoginStateRes)
         {
@@ -316,12 +334,21 @@ void UserSystem::LoginWithRefreshToken(const csp::common::String& UserId, const 
         Request->SetUserId(UserId);
         Request->SetRefreshToken(RefreshToken);
 
-        if (TokenOptions.HasValue() && CheckExpiryLengthFormat(TokenOptions->ExpiryLength))
+        auto Options = std::make_shared<chs_user::TokenOptions>();
+
+        if (TokenOptions.HasValue() && CheckExpiryLengthFormat(TokenOptions->AccessTokenExpiryLength))
         {
-            auto Options = std::make_shared<chs_user::TokenOptions>();
-            Options->SetExpiryLength(TokenOptions->ExpiryLength);
-            Request->SetTokenOptions(Options);
+            Options->SetExpiryLength(TokenOptions->AccessTokenExpiryLength);
+            CurrentLoginState.AccessTokenExpiryLength = TokenOptions->AccessTokenExpiryLength;
         }
+
+        if (TokenOptions.HasValue() && CheckExpiryLengthFormat(TokenOptions->RefreshTokenExpiryLength))
+        {
+            Options->SetRefreshTokenExpiryLength(TokenOptions->RefreshTokenExpiryLength);
+            CurrentLoginState.RefreshTokenExpiryLength = TokenOptions->RefreshTokenExpiryLength;
+        }
+
+        Request->SetTokenOptions(Options);
 
         LoginStateResultCallback LoginStateResCallback = [=](const LoginStateResult& LoginStateRes)
         {
@@ -382,12 +409,21 @@ void UserSystem::LoginAsGuest(bool CreateMultiplayerConnection, const csp::commo
             Request->SetVerifiedAgeEighteen(*UserHasVerifiedAge);
         }
 
-        if (TokenOptions.HasValue() && CheckExpiryLengthFormat(TokenOptions->ExpiryLength))
+        auto Options = std::make_shared<chs_user::TokenOptions>();
+
+        if (TokenOptions.HasValue() && CheckExpiryLengthFormat(TokenOptions->AccessTokenExpiryLength))
         {
-            auto Options = std::make_shared<chs_user::TokenOptions>();
-            Options->SetExpiryLength(TokenOptions->ExpiryLength);
-            Request->SetTokenOptions(Options);
+            Options->SetExpiryLength(TokenOptions->AccessTokenExpiryLength);
+            CurrentLoginState.AccessTokenExpiryLength = TokenOptions->AccessTokenExpiryLength;
         }
+
+        if (TokenOptions.HasValue() && CheckExpiryLengthFormat(TokenOptions->RefreshTokenExpiryLength))
+        {
+            Options->SetRefreshTokenExpiryLength(TokenOptions->RefreshTokenExpiryLength);
+            CurrentLoginState.RefreshTokenExpiryLength = TokenOptions->RefreshTokenExpiryLength;
+        }
+
+        Request->SetTokenOptions(Options);
 
         LoginStateResultCallback LoginStateResCallback = [=](const LoginStateResult& LoginStateRes)
         {
@@ -616,12 +652,21 @@ void UserSystem::LoginToThirdPartyAuthenticationProvider(const csp::common::Stri
         Request->SetVerifiedAgeEighteen(*UserHasVerifiedAge);
     }
 
-    if (TokenOptions.HasValue() && CheckExpiryLengthFormat(TokenOptions->ExpiryLength))
+    auto Options = std::make_shared<chs_user::TokenOptions>();
+
+    if (TokenOptions.HasValue() && CheckExpiryLengthFormat(TokenOptions->AccessTokenExpiryLength))
     {
-        auto Options = std::make_shared<chs_user::TokenOptions>();
-        Options->SetExpiryLength(TokenOptions->ExpiryLength);
-        Request->SetTokenOptions(Options);
+        Options->SetExpiryLength(TokenOptions->AccessTokenExpiryLength);
+        CurrentLoginState.AccessTokenExpiryLength = TokenOptions->AccessTokenExpiryLength;
     }
+
+    if (TokenOptions.HasValue() && CheckExpiryLengthFormat(TokenOptions->RefreshTokenExpiryLength))
+    {
+        Options->SetRefreshTokenExpiryLength(TokenOptions->RefreshTokenExpiryLength);
+        CurrentLoginState.RefreshTokenExpiryLength = TokenOptions->RefreshTokenExpiryLength;
+    }
+
+    Request->SetTokenOptions(Options);
 
     CurrentLoginState.State = csp::common::ELoginState::LoginRequested;